+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMHazmatWaterParser.java
@@ -24,7 +24,7 @@ public class OSMHazmatWaterParser implements TagParser {
     }
 
     @Override
-    public IntsRef handleWayTags(IntsRef edgeFlags, ReaderWay readerWay, boolean ferry, IntsRef relationFlags) {
+    public IntsRef handleWayTags(IntsRef edgeFlags, ReaderWay readerWay, IntsRef relationFlags) {
         if (readerWay.hasTag("hazmat:water", "no")) {
             hazWaterEnc.setEnum(false, edgeFlags, HazmatWater.NO);
         } else if (readerWay.hasTag("hazmat:water", "permissive")) {